requirements.txt
asgiref==3.8.1
Django==5.1.4
django-rest-framework==0.1.0
djangorestframework==3.15.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.12.2
coverage==7.6.10
djoser==2.3.1
pytest==8.3.4
pytest-django==4.9.0
